ventas/requirements.txt
fastapi
uvicorn[standard]
pytest
httpx
pytest-cov
SQLAlchemy>=2.0
psycopg2-binary
python-dotenv
google-cloud-pubsub
